+++ app/src/main/java/edu/mirea/onebeattrue/mylittlepet/domain/auth/AuthRepository.kt
@@ -1,7 +1,7 @@
 package edu.mirea.onebeattrue.mylittlepet.domain.auth
 
 import android.app.Activity
-import edu.mirea.onebeattrue.mylittlepet.domain.auth.utils.AuthState
+import edu.mirea.onebeattrue.mylittlepet.domain.auth.state.AuthState
 import kotlinx.coroutines.flow.Flow
 
 interface AuthRepository {
