+++ app/src/main/java/edu/mirea/onebeattrue/mylittlepet/presentation/screens/auth/ConfirmPhoneScreen.kt
@@ -1,6 +1,5 @@
-package edu.mirea.onebeattrue.mylittlepet.presentation.ui.screens.auth
+package edu.mirea.onebeattrue.mylittlepet.presentation.screens.auth
 
-import android.util.Log
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -14,6 +13,7 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.rounded.KeyboardArrowLeft
 import androidx.compose.material.icons.rounded.KeyboardArrowRight
 import androidx.compose.material3.Button
 import androidx.compose.material3.Card
@@ -24,12 +24,9 @@ import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.shadow
@@ -41,27 +38,26 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import edu.mirea.onebeattrue.mylittlepet.R
-import edu.mirea.onebeattrue.mylittlepet.domain.auth.utils.AuthState
-import edu.mirea.onebeattrue.mylittlepet.presentation.MainActivity
-import edu.mirea.onebeattrue.mylittlepet.presentation.ui.theme.MyLittlePetTheme
-import edu.mirea.onebeattrue.mylittlepet.presentation.viewmodels.EnterPhoneViewModel
+import edu.mirea.onebeattrue.mylittlepet.domain.auth.state.AuthState
+import edu.mirea.onebeattrue.mylittlepet.ui.theme.MyLittlePetTheme
+import edu.mirea.onebeattrue.mylittlepet.presentation.viewmodels.auth.ConfirmPhoneViewModel
 import edu.mirea.onebeattrue.mylittlepet.presentation.viewmodels.ViewModelFactory
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
 
 @Composable
-fun EnterPhoneScreen(
+fun ConfirmPhoneScreen(
     modifier: Modifier = Modifier,
-    onNextButtonClickListener: () -> Unit,
-    viewModelFactory: ViewModelFactory,
-    activity: MainActivity
+    onBackButtonClickListener: () -> Unit,
+    onConfirmButtonClickListener: () -> Unit,
+    viewModelFactory: ViewModelFactory
 ) {
-    val phoneNumber = rememberSaveable {
+    val code = rememberSaveable {
         mutableStateOf("")
     }
     val scope = rememberCoroutineScope()
-    val viewModel: EnterPhoneViewModel = viewModel(factory = viewModelFactory)
+    val viewModel: ConfirmPhoneViewModel = viewModel(factory = viewModelFactory)
 
     Column(
         modifier = modifier.fillMaxSize(),
@@ -70,7 +66,7 @@ fun EnterPhoneScreen(
     ) {
         Image(
             modifier = Modifier.size(200.dp),
-            painter = painterResource(id = R.drawable.image_cat_face),
+            painter = painterResource(id = R.drawable.image_dog_face),
             contentDescription = null,
         )
         Spacer(modifier = Modifier.height(16.dp))
@@ -102,26 +98,35 @@ fun EnterPhoneScreen(
                 horizontalAlignment = Alignment.CenterHorizontally
             ) {
                 Text(
-                    text = stringResource(id = R.string.enter_phone_number),
+                    text = stringResource(id = R.string.enter_confirmation_code),
                     fontSize = 24.sp
                 )
-                PhoneTextField(modifier = Modifier.fillMaxWidth(), phoneNumber = phoneNumber)
+                ConfirmPhoneTextField(modifier = Modifier.fillMaxWidth(), code = code)
                 Row(
                     modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.End
+                    horizontalArrangement = Arrangement.SpaceBetween
                 ) {
+                    Button(
+                        onClick = { onBackButtonClickListener() },
+                        shape = RoundedCornerShape(16.dp)
+                    ) {
+                        Icon(
+                            imageVector = Icons.Rounded.KeyboardArrowLeft,
+                            contentDescription = null
+                        )
+                        Text(
+                            text = stringResource(id = R.string.back),
+                            fontSize = 16.sp
+                        )
+                    }
                     Button(
                         onClick = {
                             scope.launch(Dispatchers.Main) {
-                                viewModel.createUserWithPhone(phoneNumber.value, activity).collect {
+                                viewModel.signUpWithCredential(code.value).collect {
                                     when (it) {
-                                        is AuthState.Failure -> {
-                                            Log.d("EnterPhoneScreen", it.exception.message.toString())
-                                        }
-                                        is AuthState.Success -> onNextButtonClickListener()
-                                        AuthState.Loading -> {
-
-                                        }
+                                        is AuthState.Failure -> TODO()
+                                        is AuthState.Success -> onConfirmButtonClickListener()
+                                        AuthState.Loading -> TODO()
                                     }
                                 }
                             }
@@ -129,7 +134,7 @@ fun EnterPhoneScreen(
                         shape = RoundedCornerShape(16.dp)
                     ) {
                         Text(
-                            text = stringResource(id = R.string.next),
+                            text = stringResource(id = R.string.confirm),
                             fontSize = 16.sp
                         )
                         Icon(
@@ -143,37 +148,46 @@ fun EnterPhoneScreen(
     }
 }
 
+
 @Composable
-private fun PhoneTextField(
+private fun ConfirmPhoneTextField(
     modifier: Modifier = Modifier,
-    phoneNumber: MutableState<String>
+    code: MutableState<String>
 ) {
     OutlinedTextField(
         modifier = modifier,
-        value = phoneNumber.value,
-        onValueChange = { phoneNumber.value = it },
+        value = code.value,
+        onValueChange = { code.value = it },
         label = {
-            Text(stringResource(id = R.string.phone_number_hint))
+            Text(stringResource(id = R.string.confirmation_code_hint))
         },
-        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Phone),
+        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
         shape = RoundedCornerShape(16.dp),
         singleLine = true,
     )
 }
 
 
-//@Preview
-//@Composable
-//private fun EnterPhoneScreenPreviewLight() {
-//    MyLittlePetTheme(darkTheme = false) {
-//        EnterPhoneScreen()
-//    }
-//}
-//
-//@Preview
-//@Composable
-//private fun EnterPhoneScreenPreviewDark() {
-//    MyLittlePetTheme(darkTheme = true) {
-//        EnterPhoneScreen()
-//    }
-//}
+@Preview
+@Composable
+private fun ConfirmPhoneScreenPreviewLight() {
+    MyLittlePetTheme(darkTheme = false) {
+        ConfirmPhoneScreen(
+            onBackButtonClickListener = {},
+            onConfirmButtonClickListener = {},
+            viewModelFactory = ViewModelFactory(mapOf())
+        )
+    }
+}
+
+@Preview
+@Composable
+private fun ConfirmPhoneScreenPreviewDark() {
+    MyLittlePetTheme(darkTheme = true) {
+        ConfirmPhoneScreen(
+            onBackButtonClickListener = {},
+            onConfirmButtonClickListener = {},
+            viewModelFactory = ViewModelFactory(mapOf())
+        )
+    }
+}
