+++ app/src/main/java/edu/mirea/onebeattrue/mylittlepet/presentation/screens/auth/EnterPhoneScreen.kt
@@ -1,5 +1,6 @@
-package edu.mirea.onebeattrue.mylittlepet.presentation.ui.screens.auth
+package edu.mirea.onebeattrue.mylittlepet.presentation.screens.auth
 
+import android.util.Log
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -13,55 +14,49 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.rounded.KeyboardArrowLeft
 import androidx.compose.material.icons.rounded.KeyboardArrowRight
 import androidx.compose.material3.Button
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import edu.mirea.onebeattrue.mylittlepet.R
-import edu.mirea.onebeattrue.mylittlepet.domain.auth.utils.AuthState
+import edu.mirea.onebeattrue.mylittlepet.domain.auth.state.AuthState
 import edu.mirea.onebeattrue.mylittlepet.presentation.MainActivity
-import edu.mirea.onebeattrue.mylittlepet.presentation.ui.theme.MyLittlePetTheme
-import edu.mirea.onebeattrue.mylittlepet.presentation.viewmodels.ConfirmPhoneViewModel
+import edu.mirea.onebeattrue.mylittlepet.presentation.viewmodels.auth.EnterPhoneViewModel
 import edu.mirea.onebeattrue.mylittlepet.presentation.viewmodels.ViewModelFactory
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
 
 @Composable
-fun ConfirmPhoneScreen(
+fun EnterPhoneScreen(
     modifier: Modifier = Modifier,
-    onBackButtonClickListener: () -> Unit,
-    onConfirmButtonClickListener: () -> Unit,
-    viewModelFactory: ViewModelFactory
+    onNextButtonClickListener: () -> Unit,
+    viewModelFactory: ViewModelFactory,
+    activity: MainActivity
 ) {
-    val code = rememberSaveable {
+    val phoneNumber = rememberSaveable {
         mutableStateOf("")
     }
     val scope = rememberCoroutineScope()
-    val viewModel: ConfirmPhoneViewModel = viewModel(factory = viewModelFactory)
+    val viewModel: EnterPhoneViewModel = viewModel(factory = viewModelFactory)
 
     Column(
         modifier = modifier.fillMaxSize(),
@@ -70,7 +65,7 @@ fun ConfirmPhoneScreen(
     ) {
         Image(
             modifier = Modifier.size(200.dp),
-            painter = painterResource(id = R.drawable.image_dog_face),
+            painter = painterResource(id = R.drawable.image_cat_face),
             contentDescription = null,
         )
         Spacer(modifier = Modifier.height(16.dp))
@@ -102,35 +97,26 @@ fun ConfirmPhoneScreen(
                 horizontalAlignment = Alignment.CenterHorizontally
             ) {
                 Text(
-                    text = stringResource(id = R.string.enter_confirmation_code),
+                    text = stringResource(id = R.string.enter_phone_number),
                     fontSize = 24.sp
                 )
-                ConfirmPhoneTextField(modifier = Modifier.fillMaxWidth(), code = code)
+                PhoneTextField(modifier = Modifier.fillMaxWidth(), phoneNumber = phoneNumber)
                 Row(
                     modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.SpaceBetween
+                    horizontalArrangement = Arrangement.End
                 ) {
-                    Button(
-                        onClick = { onBackButtonClickListener() },
-                        shape = RoundedCornerShape(16.dp)
-                    ) {
-                        Icon(
-                            imageVector = Icons.Rounded.KeyboardArrowLeft,
-                            contentDescription = null
-                        )
-                        Text(
-                            text = stringResource(id = R.string.back),
-                            fontSize = 16.sp
-                        )
-                    }
                     Button(
                         onClick = {
                             scope.launch(Dispatchers.Main) {
-                                viewModel.signUpWithCredential(code.value).collect {
+                                viewModel.createUserWithPhone(phoneNumber.value, activity).collect {
                                     when (it) {
-                                        is AuthState.Failure -> TODO()
-                                        is AuthState.Success -> onConfirmButtonClickListener()
-                                        AuthState.Loading -> TODO()
+                                        is AuthState.Failure -> {
+                                            Log.d("EnterPhoneScreen", it.exception.message.toString())
+                                        }
+                                        is AuthState.Success -> onNextButtonClickListener()
+                                        AuthState.Loading -> {
+
+                                        }
                                     }
                                 }
                             }
@@ -138,7 +124,7 @@ fun ConfirmPhoneScreen(
                         shape = RoundedCornerShape(16.dp)
                     ) {
                         Text(
-                            text = stringResource(id = R.string.confirm),
+                            text = stringResource(id = R.string.next),
                             fontSize = 16.sp
                         )
                         Icon(
@@ -152,46 +138,37 @@ fun ConfirmPhoneScreen(
     }
 }
 
-
 @Composable
-private fun ConfirmPhoneTextField(
+private fun PhoneTextField(
     modifier: Modifier = Modifier,
-    code: MutableState<String>
+    phoneNumber: MutableState<String>
 ) {
     OutlinedTextField(
         modifier = modifier,
-        value = code.value,
-        onValueChange = { code.value = it },
+        value = phoneNumber.value,
+        onValueChange = { phoneNumber.value = it },
         label = {
-            Text(stringResource(id = R.string.confirmation_code_hint))
+            Text(stringResource(id = R.string.phone_number_hint))
         },
-        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
+        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Phone),
         shape = RoundedCornerShape(16.dp),
         singleLine = true,
     )
 }
 
 
-@Preview
-@Composable
-private fun ConfirmPhoneScreenPreviewLight() {
-    MyLittlePetTheme(darkTheme = false) {
-        ConfirmPhoneScreen(
-            onBackButtonClickListener = {},
-            onConfirmButtonClickListener = {},
-            viewModelFactory = ViewModelFactory(mapOf())
-        )
-    }
-}
-
-@Preview
-@Composable
-private fun ConfirmPhoneScreenPreviewDark() {
-    MyLittlePetTheme(darkTheme = true) {
-        ConfirmPhoneScreen(
-            onBackButtonClickListener = {},
-            onConfirmButtonClickListener = {},
-            viewModelFactory = ViewModelFactory(mapOf())
-        )
-    }
-}
+//@Preview
+//@Composable
+//private fun EnterPhoneScreenPreviewLight() {
+//    MyLittlePetTheme(darkTheme = false) {
+//        EnterPhoneScreen()
+//    }
+//}
+//
+//@Preview
+//@Composable
+//private fun EnterPhoneScreenPreviewDark() {
+//    MyLittlePetTheme(darkTheme = true) {
+//        EnterPhoneScreen()
+//    }
+//}
