+++ app/src/main/java/edu/mirea/onebeattrue/mylittlepet/presentation/screens/main/ProfileScreen.kt
@@ -1,4 +1,4 @@
-package edu.mirea.onebeattrue.mylittlepet.presentation.ui.screens.main
+package edu.mirea.onebeattrue.mylittlepet.presentation.screens.main
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
