--- app/src/main/java/edu/mirea/onebeattrue/mylittlepet/presentation/ui/screens/auth/AuthScreenState.kt
@@ -1,8 +1,0 @@
-package edu.mirea.onebeattrue.mylittlepet.presentation.ui.screens.auth
-
-sealed class AuthScreenState {
-    object Initial : AuthScreenState()
-    data class EnterPhone(val phoneNumber: String) : AuthScreenState()
-    data class ConfirmPhone(val code: Int) : AuthScreenState()
-    object Failure : AuthScreenState()
-}
