+++ app/src/main/java/edu/mirea/onebeattrue/mylittlepet/presentation/viewmodels/auth/ConfirmPhoneViewModel.kt
@@ -1,4 +1,4 @@
-package edu.mirea.onebeattrue.mylittlepet.presentation.viewmodels
+package edu.mirea.onebeattrue.mylittlepet.presentation.viewmodels.auth
 
 import androidx.lifecycle.ViewModel
 import edu.mirea.onebeattrue.mylittlepet.domain.auth.AuthRepository
