+++ app/src/main/java/edu/mirea/onebeattrue/mylittlepet/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package edu.mirea.onebeattrue.mylittlepet.presentation.ui.theme
+package edu.mirea.onebeattrue.mylittlepet.ui.theme
 
 import androidx.compose.ui.graphics.Color
